First stage in upgrade to parallel R 2.15.1.
git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@1329 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/library/Recommended/nlme/src/gnls.c
@@ -57,7 +57,7 @@ gnls_init(double *ptheta, longint *dims, double *corFactor, double *varWeights,
   gnls->N = dims[1];
   gnls->nrdof = gnls->N - gnls->npar;
   gnls->ncol = gnls->npar + 1;
-  gnls->maxIter = (long) settings[0];
+  gnls->maxIter = (int) settings[0];
   gnls->minFactor = settings[1];
   gnls->tolerance = settings[2];
   gnls->newtheta = Calloc(gnls->npar, double);
